# openapi.yaml
openapi: 3.1.0
info:
  description: LensBridge backend — media upload and MusallahBoard management
  title: LensBridge API
  version: 0.0.1-SNAPSHOT
servers:
- url: /
security:
- bearerAuth: []
paths:
...
  /api/admin/board/devices:
    get:
      operationId: listDevices
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeviceSummary"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      summary: List every enrolled board device
      tags:
      - device-admin-controller
  /api/admin/board/devices/enrollment-tokens:
    post:
      operationId: issueEnrollmentToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueEnrollmentTokenRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueEnrollmentTokenResponse"
          description: Token issued; the plaintext is returned exactly once
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      summary: Mint a one-time token a new device can exchange for an identity
      tags:
      - device-admin-controller
  /api/admin/board/devices/{deviceId}:
    get:
      operationId: getDevice
      parameters:
      - in: path
        name: deviceId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceSummary"
          description: Device found
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: No device with that id
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      summary: Fetch a single board device
      tags:
      - device-admin-controller
    patch:
      operationId: updateDevice
      parameters:
      - in: path
        name: deviceId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeviceRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceSummary"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      tags:
      - board-admin-controller
  /api/admin/board/devices/{deviceId}/commands:
    get:
      operationId: listDeviceCommands
      parameters:
      - in: path
        name: deviceId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommandView"
          description: OK
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      summary: Fetch the 50 most recent commands for a device
      tags:
      - device-admin-controller
    post:
      description: "Returns as soon as the command is queued; delivery and execution\
        \ are reported asynchronously over the agent websocket. The permission required\
        \ depends on the command kind: board:command:benign, :disruptive, or :inspect."
      operationId: issueDeviceCommand
      parameters:
      - in: path
        name: deviceId
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IssueCommandRequest"
        required: true
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandIssuedResponse"
          description: Command accepted and queued for delivery
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Caller lacks the permission for this command kind
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      summary: Queue a command for a device
      tags:
      - device-admin-controller
  /api/admin/board/devices/{deviceId}/revoke:
    post:
      description: "Closes any live agent session. Idempotent: re-revoking an already\
        \ revoked device is a no-op."
      operationId: revokeDevice
      parameters:
      - in: path
        name: deviceId
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceSummary"
          description: Device revoked
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: No device with that id
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      summary: Revoke a device's enrollment
      tags:
      - device-admin-controller
...
components:
  schemas:
...
    CommandIssuedResponse:
      type: object
      properties:
        commandId:
          type: string
          format: uuid
        deviceId:
          type: string
          format: uuid
        expiresAt:
          type: string
          format: date-time
        issuedAt:
          type: string
          format: date-time
        kind:
          type: string
        status:
          type: string
          enum:
          - PENDING
          - DELIVERED
          - ACKED
          - RUNNING
          - SUCCEEDED
          - FAILED
          - TIMEOUT
          - REJECTED
          - EXPIRED
    CommandView:
      type: object
      properties:
        ackedAt:
          type: string
          format: date-time
        deadlineMs:
          type: integer
          format: int32
        deliveredAt:
          type: string
          format: date-time
        deviceId:
          type: string
          format: uuid
        errorMessage:
          type: string
        expiresAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        id:
          type: string
          format: uuid
        issuedAt:
          type: string
          format: date-time
        issuedBy:
          type: string
        kind:
          type: string
        output:
          $ref: "#/components/schemas/JsonNode"
        payload:
          $ref: "#/components/schemas/JsonNode"
        startedAt:
          type: string
          format: date-time
        status:
          type: string
          enum:
          - PENDING
          - DELIVERED
          - ACKED
          - RUNNING
          - SUCCEEDED
          - FAILED
          - TIMEOUT
          - REJECTED
          - EXPIRED
...
    DeviceSummary:
      type: object
      properties:
        agentVersion:
          type: string
        audience:
          type: string
          enum:
          - brothers
          - sisters
          - both
        displayName:
          type: string
        enrolledAt:
          type: string
          format: date-time
        hardwareModel:
          type: string
        id:
          type: string
          format: uuid
        lastHeartbeat:
          type: string
          format: date-time
        lastSeenIp:
          type: string
        revokedAt:
          type: string
          format: date-time
...
    IssueCommandRequest:
      type: object
      properties:
        deadlineMs:
          type: integer
          format: int32
          minimum: 0
        kind:
          type: string
          minLength: 1
          pattern: "[a-z]+\\.[a-z_]+"
        payload:
          $ref: "#/components/schemas/JsonNode"
        ttlSeconds:
          type: integer
          format: int32
          minimum: 0
      required:
      - kind
    IssueEnrollmentTokenRequest:
      type: object
      properties:
        audience:
          type: string
          enum:
          - brothers
          - sisters
          - both
        displayName:
          type: string
          minLength: 1
        expiresInMinutes:
          type: integer
          format: int32
          minimum: 0
      required:
      - audience
      - displayName
    IssueEnrollmentTokenResponse:
      type: object
      properties:
        expiresAt:
          type: string
          format: date-time
        token:
          type: string
        tokenId:
          type: string
          format: uuid
    JsonNode: {}
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    UpdateDeviceRequest:
      type: object
      properties:
        audience:
          type: string
          enum:
          - brothers
          - sisters
          - both
        displayName:
          type: string
          maxLength: 255
          minLength: 0
          pattern: .*\S.*
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      name: bearerAuth
      scheme: bearer
      type: http